+++ src/lib.rs
@@ -1,1 +1,7 @@
-pub mod prelude {}
+mod task;
+mod traits;
+pub mod prelude {
+    use super::*;
+    pub use task::{Task, TaskGraph};
+    pub use traits::{Execute, Register};
+}
